Moving ARRF implementation into AggregateOf and removing ARRF
